Add multi line commit message
I am adding a multiline commit message to test my understanding of how
commit messages work. As of today, the rules I remember include writing
the commit title in the "imperative" form. Another rule is that text
warps after rougly 70 characters so I should be mindful of that. Another
rule is that explain the what and why of the changes, not the how. I can
 always look at the github lesson for more details.